Decouple parsing and formatting, add regexp-based parsing.
Changes parse_*() to format_*() (since they format internal deck representations into user-viewable strings) and adds replacement parse_*() functions that are intended to actually validate arbitrary user input. They use regexps and do basic sanity checks to ensure that input is sensible; if it is not, they raise a ParseError which can be caught by the calling code to, say, pass it along to the user so they know why it didn't work.