Fix comparison of string to integer. Fix reversed suit definitions.
`parse_hanafuda()` was incorrectly comparing `suit` (a string) to tuples of numbers, which meant the final value check could not validate q/p/t correctly, and would always generate a `ParseError()` when those values were used; in other words, 50% of the cards in a Hanafuda deck were broken.

Additionally, the regular expressions with the duplicated suit definitions (to handle both "Hearts Ace" and "Ace of Hearts") actually did not work because of the duplicate definition of "suit", despite only one definition being able to be filled at a time. I'm not sure how I didn't come across this problem when designing the regexps in the first place.

Additionally, while `parse_card()` disallowed "Joker of Spades", a suitless "Jack" would erroneously parse as a Joker, and no validation of suitless cards was made.

And lastly, despite `suit` sometimes being `None`, we tried to call `.casefold()` on it unconditionally.